requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
coverage==7.4.1
cryptography==42.0.3
exceptiongroup==1.2.0
fastapi==0.109.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
Mako==1.3.2
MarkupSafe==2.1.5
numpy==1.26.2
packaging==23.2
pluggy==1.4.0
psycopg-binary==3.1.18
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.8.0
pytest==8.0.0
pytest-cov==4.1.0
pytest-dotenv==0.5.2
python-dotenv==1.0.1
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.36.3
tomli==2.0.1
typing_extensions==4.9.0
uvicorn==0.27.0.post1
